+++ src/writer/junit.rs
@@ -25,7 +25,7 @@ use crate::{
         self,
         basic::{coerce_error, Coloring},
         out::WritableString,
-        Ext as _, Repeatable,
+        Ext as _,
     },
     Event, World, Writer,
 };
@@ -133,7 +133,7 @@ where
     }
 }
 
-impl<W, O: io::Write> Repeatable for JUnit<W, O> {}
+impl<W, O: io::Write> writer::NotTransformEvents for JUnit<W, O> {}
 
 impl<W: Debug + World, Out: io::Write> JUnit<W, Out> {
     /// Creates a new normalized [`JUnit`] [`Writer`] outputting XML report into